adding bump, optimize, test
