+++ gulpfile.js
@@ -49,6 +49,9 @@ require('./gulp-tasks/assets')(env);
 require('./gulp-tasks/clean')(env);
 require('./gulp-tasks/template-cache')(env);
 require('./gulp-tasks/inject')(env);
+require('./gulp-tasks/optimize')(env);
+require('./gulp-tasks/test')(env);
+require('./gulp-tasks/bump')(env);
 
 
 gulp.task('default', [ 'help' ]);
@@ -64,15 +67,6 @@ gulp.task('default', [ 'help' ]);
  * --startServers: Will start servers for midway tests on the test task.
  */
 
-/**
- * Create $templateCache from the html templates
- * @return {Stream}
- */
-
-/**
- * Wire-up the bower dependencies
- * @return {Stream}
- */
 /**
  * Run the spec runner
  * @return {Stream}
@@ -126,72 +120,6 @@ gulp.task('build', [ 'optimize', 'images', 'fonts' ], function () {
 });
 
 /**
- * Optimize all files, move to a build folder,
- * and inject them into the new index.html
- * @return {Stream}
- */
-gulp.task('optimize', [ 'inject', 'test' ], function () {
-    log('Optimizing the js, css, and html');
-
-    var assets = $.useref.assets({ searchPath: './' });
-    // Filters are named for the gulp-useref path
-    var cssFilter = $.filter('**/*.css');
-    var jsAppFilter = $.filter('**/' + config.optimized.app);
-    var jslibFilter = $.filter('**/' + config.optimized.lib);
-
-    var templateCache = config.temp + config.templateCache.file;
-
-    return gulp
-        .src(config.index)
-        .pipe($.plumber())
-        .pipe($.inject(gulp.src(templateCache),
-            { name: 'inject:templates', read: false }))
-        .pipe(assets) // Gather all assets from the html with useref
-        // Get the css
-        .pipe(cssFilter)
-        .pipe($.csso())
-        .pipe(cssFilter.restore())
-        // Get the custom javascript
-        .pipe(jsAppFilter)
-        .pipe($.ngAnnotate({ add: true }))
-        .pipe($.uglify())
-        .pipe(getHeader())
-        .pipe(jsAppFilter.restore())
-        // Get the vendor javascript
-        .pipe(jslibFilter)
-        .pipe($.uglify()) // another option is to override wiredep to use min files
-        .pipe(jslibFilter.restore())
-        // Take inventory of the file names for future rev numbers
-        .pipe($.rev())
-        // Apply the concat and file replacement with useref
-        .pipe(assets.restore())
-        .pipe($.useref())
-        // Replace the file names in the html with rev numbers
-        .pipe($.revReplace())
-        .pipe(gulp.dest(config.build));
-});
-
-/**
- * Run specs once and exit
- * To start servers and run midway specs as well:
- *    gulp test --startServers
- * @return {Stream}
- */
-gulp.task('test', [ 'vet', 'templatecache' ], function (done) {
-    startTests(true /*singleRun*/, done);
-});
-
-/**
- * Run specs and wait.
- * Watch for file changes and re-run tests on each change
- * To start servers and run midway specs as well:
- *    gulp autotest --startServers
- */
-gulp.task('autotest', function (done) {
-    startTests(false /*singleRun*/, done);
-});
-
-/**
  * serve the dev environment
  * --debug-brk or --debug
  * --nosync
@@ -216,34 +144,6 @@ gulp.task('serve-build', [ 'build' ], function () {
     serve(false /*isDev*/);
 });
 
-/**
- * Bump the version
- * --type=pre will bump the prerelease version *.*.*-x
- * --type=patch or no flag will bump the patch version *.*.x
- * --type=minor will bump the minor version *.x.*
- * --type=major will bump the major version x.*.*
- * --version=1.2.3 will bump to a specific version and ignore other flags
- */
-gulp.task('bump', function () {
-    var msg = 'Bumping versions';
-    var type = args.type;
-    var version = args.ver;
-    var options = {};
-    if (version) {
-        options.version = version;
-        msg += ' to ' + version;
-    } else {
-        options.type = type;
-        msg += ' for a ' + type;
-    }
-    log(msg);
-
-    return gulp
-        .src(config.packages)
-        .pipe($.print())
-        .pipe($.bump(options))
-        .pipe(gulp.dest(config.root));
-});
 
 ////////////////
 
@@ -272,12 +172,6 @@ function changeEvent(event) {
     var srcPattern = new RegExp('/.*(?=/' + config.source + ')/');
     log('File ' + event.path.replace(srcPattern, '') + ' ' + event.type);
 }
-
-/**
- * Delete all files in a given path
- * @param  {Array}   path - array of paths to delete
- * @param  {Function} done - callback when complete
- */
 
 /**
  * serve the code
@@ -377,59 +271,6 @@ function startBrowserSync(isDev, specRunner) {
 }
 
 /**
- * Start the tests using karma.
- * @param  {boolean} singleRun - True means run once and end (CI), or keep running (dev)
- * @param  {Function} done - Callback to fire when karma is done
- * @return {undefined}
- */
-function startTests(singleRun, done) {
-    var child;
-    var excludeFiles = [];
-    var fork = require('child_process').fork;
-    var karma = require('karma').server;
-    var serverSpecs = config.serverIntegrationSpecs;
-
-    if (args.startServers) {
-        log('Starting servers');
-        var savedEnv = process.env;
-        savedEnv.NODE_ENV = 'dev';
-        savedEnv.PORT = 3000;
-        child = fork(config.nodeServer);
-    } else {
-        if (serverSpecs && serverSpecs.length) {
-            excludeFiles = serverSpecs;
-        }
-    }
-
-    karma.start({
-        configFile: __dirname + '/karma.conf.js',
-        exclude: excludeFiles,
-        singleRun: !!singleRun
-    }, karmaCompleted);
-
-    ////////////////
-
-    function karmaCompleted(karmaResult) {
-        log('Karma completed');
-        if (child) {
-            log('shutting down the child process');
-            child.kill();
-        }
-        if (karmaResult === 1) {
-            done('karma: tests failed with code ' + karmaResult);
-        } else {
-            done();
-        }
-    }
-}
-
-/**
- * Formatter for bytediff to display the size changes after processing
- * @param  {Object} data - byte data
- * @return {String}      Difference in bytes, formatted
- */
-
-/**
  * Log an error message and emit the end of a task
  */
 function errorLogger(error) {
@@ -437,33 +278,6 @@ function errorLogger(error) {
     log(error);
     log('*** End of Error ***');
     this.emit('end');
-}
-
-/**
- * Format a number as a percentage
- * @param  {Number} num       Number to format as a percent
- * @param  {Number} precision Precision of the decimal
- * @return {String}           Formatted perentage
- */
-
-/**
- * Format and return the header for files
- * @return {String}           Formatted file header
- */
-function getHeader() {
-    var pkg = require('./package.json');
-    var template = [ '/**',
-        ' * <%= pkg.name %> - <%= pkg.description %>',
-        ' * @authors <%= pkg.authors %>',
-        ' * @version v<%= pkg.version %>',
-        ' * @link <%= pkg.homepage %>',
-        ' * @license <%= pkg.license %>',
-        ' */',
-        ''
-    ].join('\n');
-    return $.header(template, {
-        pkg: pkg
-    });
 }
 
 /**
